Website-smarthome/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
